Moves command handling to an event handler, py3 support
Commands are now handled by capturing privmsg events in the bot event
handler and handling that logic in a sub function. This means there is
no longer an exception in the logic method for PRIVMSG events, and to
reflect that it was renamed from logic to handle_event

The code was also updated to add python3 support

Some basic skeletal work was done on adding ctcp support, mostly the
inclusion of the new irc wrapper methods notice and notice_all which can
be used to send notices.